Cleans up reference model
