+++ models/experimental/detr3d/reference/transformer.py
@@ -1,12 +1,17 @@
-# SPDX-FileCopyrightText: © 2025 Tenstorrent AI ULC
-# SPDX-License-Identifier: Apache-2.0
-# Modified from the 3Detr implementation in facebookresearch
-# https://github.com/facebookresearch/3detr
 # Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
+"""
+Modified from DETR Transformer class.
+
+Copy-paste from torch.nn.Transformer with modifications:
+    * positional encodings are passed in MHattention
+    * extra LN at the end of encoder is removed
+    * decoder returns a stack of activations from all decoding layers
+"""
+from typing import Optional
 
 import torch
-from typing import Optional
 from torch import Tensor, nn
+
 from models.experimental.detr3d.reference.helpers import ACTIVATION_DICT, NORM_DICT, WEIGHT_INIT_DICT, get_clones
