+++ models/experimental/detr3d/tests/test_transformer_encoder_layer.py
@@ -5,7 +5,7 @@
 import torch
 import pytest
 from models.experimental.detr3d.source.detr3d.models.transformer import TransformerEncoderLayer as org_model
-from models.experimental.detr3d.reference.detr3d_model import TransformerEncoderLayer as ref_model
+from models.experimental.detr3d.reference.model_3detr import TransformerEncoderLayer as ref_model
 from tests.ttnn.utils_for_testing import assert_with_pcc
